+++ packages/elements-core/src/components/TableOfContents/types.ts
@@ -1,3 +1,5 @@
+import type { TableOfContentsItem } from '@stoplight/elements-utils';
+
 export type TableOfContentsProps = {
   tree: TableOfContentsItem[];
   activeId: string;
@@ -14,38 +16,12 @@ export type CustomLinkComponent = React.ComponentType<{
   children: React.ReactNode;
 }>;
 
-export type TableOfContentsItem = TableOfContentsDivider | TableOfContentsGroupItem;
-
-export type TableOfContentsDivider = {
-  title: string;
-};
-
-export type TableOfContentsGroupItem =
-  | TableOfContentsGroup
-  | TableOfContentsNodeGroup
-  | TableOfContentsNode
-  | TableOfContentsExternalLink;
-
-export type TableOfContentsGroup = {
-  title: string;
-  items: TableOfContentsGroupItem[];
-  itemsType?: 'article' | 'http_operation' | 'http_webhook' | 'model';
-};
-
-export type TableOfContentsExternalLink = {
-  title: string;
-  url: string;
-};
-
-export type TableOfContentsNode<
-  T = 'http_service' | 'http_operation' | 'http_webhook' | 'model' | 'article' | 'overview',
-> = {
-  id: string;
-  slug: string;
-  title: string;
-  type: T;
-  meta: string;
-  version?: string;
-};
-
-export type TableOfContentsNodeGroup = TableOfContentsNode<'http_service'> & TableOfContentsGroup;
+export type {
+  TableOfContentsDivider,
+  TableOfContentsExternalLink,
+  TableOfContentsGroup,
+  TableOfContentsGroupItem,
+  TableOfContentsItem,
+  TableOfContentsNode,
+  TableOfContentsNodeGroup,
+} from '@stoplight/elements-utils';
