+++ packages/elements-core/src/utils/string.ts
@@ -1,11 +1,5 @@
-import { curry } from 'lodash';
+export {
+  caseInsensitivelyEquals,
+  slugify
+} from '@stoplight/elements-utils';
 
-export const caseInsensitivelyEquals = curry((a: string, b: string) => a.toUpperCase() === b.toUpperCase());
-
-export function slugify(name: string) {
-  return name
-    .replace(/\/|{|}|\s/g, '-')
-    .replace(/-{2,}/, '-')
-    .replace(/^-/, '')
-    .replace(/-$/, '');
-}
