+++ packages/elements-dev-portal/src/version.ts
@@ -1,2 +1,2 @@
 // auto-updated during build
-export const appVersion = '2.4.2';
+export const appVersion = '2.5.2';
