+++ packages/elements-utils/README.md
@@ -1,0 +1,7 @@
+# Elements utils
+
+## Code to extract table of contents from ACT generated by `http-spec`
+
+## COde to provide readable descriptions for security sections extracted from ACT generated by `http-spec`
+
+
