+++ packages/elements-utils/jest.config.js
@@ -1,0 +1,10 @@
+module.exports = {
+  ...require('../../jest.config'),
+  reporters: [
+    'default',
+    [
+      'jest-junit',
+      { suiteName: 'elements-utils', outputFile: '<rootDir>/../../test-results/elements-utils/results.xml' },
+    ],
+  ],
+};
