+++ packages/elements-utils/src/tableOfContents/__tests__/utils.test.ts
@@ -1,9 +1,9 @@
 import { NodeType } from '@stoplight/types';
 import { OpenAPIObject as _OpenAPIObject, PathObject } from 'openapi3-ts';
 
-import { transformOasToServiceNode } from '../../../utils/oas';
-import { OperationNode, SchemaNode, WebhookNode } from '../../../utils/oas/types';
-import {  computeTagGroups } from '../utils';
+import { transformOasToServiceNode } from '../../oas';
+import { OperationNode, SchemaNode, WebhookNode } from '../../oas/types';
+import { computeAPITree, computeTagGroups } from '../utils';
 
 type OpenAPIObject = Partial<_OpenAPIObject> & {
   webhooks?: PathObject;
